+++ Elements/src/Validators/Validators.cs
@@ -339,9 +339,11 @@ namespace Elements.Validators
 
         private void UpdateGeometry(Sweep sweep)
         {
-            var profileTrans = new Transform();
-            profileTrans.Rotate(profileTrans.ZAxis, sweep.ProfileRotation);
-            sweep._solid = Kernel.Instance.CreateSweepAlongCurve(profileTrans.OfProfile(sweep.Profile), sweep.Curve, sweep.StartSetback, sweep.EndSetback);
+            sweep._solid = Kernel.Instance.CreateSweepAlongCurve(sweep.Profile,
+                                                                 sweep.Curve,
+                                                                 sweep.StartSetback,
+                                                                 sweep.EndSetback,
+                                                                 sweep.ProfileRotation);
         }
 
         public void PreConstruct(object[] args)